[Cache] Add PdoTagAwareAdapter
Tags are stored in their own table, one (item, tag) row each, so saving,
invalidating and pruning are plain relational statements instead of the
extra round trips TagAwareAdapter needs on top of PdoAdapter. The shared
PDO plumbing moves from PdoAdapter into a new PdoTrait, and TAGS_PREFIX
becomes overridable so the adapter can store bare tag names: with a
dedicated table there is no shared key space to disambiguate.
