+++ src/Symfony/Component/Cache/Tests/Adapter/PdoTagAwareAdapterTest.php
@@ -14,14 +14,16 @@ namespace Symfony\Component\Cache\Tests\Adapter;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\RequiresPhpExtension;
 use Psr\Cache\CacheItemPoolInterface;
-use Symfony\Component\Cache\Adapter\PdoAdapter;
+use Symfony\Component\Cache\Adapter\PdoTagAwareAdapter;
 
 #[RequiresPhpExtension('pdo_sqlite')]
 #[Group('time-sensitive')]
-class PdoAdapterTest extends AbstractPdoAdapterTestCase
+class PdoTagAwareAdapterTest extends AbstractPdoAdapterTestCase
 {
+    use TagAwareTestTrait;
+
     public function createCachePool(int $defaultLifetime = 0): CacheItemPoolInterface
     {
-        return new PdoAdapter('sqlite:'.self::$dbFile, 'ns', $defaultLifetime);
+        return new PdoTagAwareAdapter('sqlite:'.self::$dbFile, 'ns', $defaultLifetime);
     }
 }
